refactor task metrics
