+++ packages/datatrak-web/src/components/index.ts
@@ -19,4 +19,3 @@ export { TextInput } from './TextInput';
 export { Tile, LoadingTile } from './Tile';
 export { Toast } from './Toast';
 export { TopProgressBar } from './TopProgressBar';
-export { TaskMetrics } from './TaskMetrics';
