+++ packages/datatrak-web/src/views/Tasks/TasksDashboardPage.tsx
@@ -9,7 +9,7 @@ import { Add } from '@material-ui/icons';
 import { Button } from '../../components';
 import { CreateTaskModal, TaskPageHeader, TasksTable } from '../../features';
 import { TasksContentWrapper } from '../../layout';
-import { TaskMetrics } from '../../components/TaskMetrics';
+import { TaskMetrics } from '../../features/Tasks/TaskMetrics';
 
 const ButtonContainer = styled.div`
   padding-block-end: 0.5rem;
